feat: se añade modulo de usuario

+++ app/Services/User/UserService.php
@@ -3,22 +3,26 @@
 namespace App\Services\User;
 
 use App\Models\User;
-use Illuminate\Support\Str;
+use App\Mail\User\UserMail;
+use App\Services\File\FileService;
 use Illuminate\Support\Facades\DB;
-use Spatie\Permission\Models\Role;
 use Illuminate\Support\Facades\Log;
-use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Mail;
-use Illuminate\Support\Facades\Password;
+use App\Interfaces\Role\RoleRepositoryInterface;
 use App\Interfaces\User\UserRepositoryInterface;
 
 class UserService
 {
     protected $userRepository;
+    protected $roleRepository;
+    protected $fileService;
+    private $listPhotos = ['photo'];
 
-    public function __construct(UserRepositoryInterface $userRepository)
+    public function __construct(UserRepositoryInterface $userRepository, RoleRepositoryInterface $roleRepository, FileService $fileService)
     {
+        $this->fileService = $fileService;
         $this->userRepository = $userRepository;
+        $this->roleRepository = $roleRepository;
     }
 
     public function getUsersQuery($role = null)
@@ -30,6 +34,19 @@ class UserService
                 $query->where('name', $role);
             }
         });
+
+        $consult->leftJoin('user_properties as up', 'up.user_id', '=', 'users.id')
+            ->leftJoin('properties as p', 'p.id', '=', 'up.property_id')
+            ->groupBy([
+                'users.id',
+                'users.name',
+                'users.phone',
+                'users.code_number',
+                'users.code_country',
+                'users.email',
+                'users.photo',
+            ]);
+
         return $consult;
     }
 
@@ -37,10 +54,27 @@ class UserService
     {
         DB::beginTransaction();
         try {
+            $photos = $data['photos'];
+            unset($data['photos']);
             $user = $this->userRepository->create($data);
+            $role = $this->roleRepository->findById($data['role']);
+            $this->assignedPhoto($user, $photos);
+
+            if ($role && $user) {
+                $user->assignRole($role);
+                if ($user->save()) {
+                    $this->sendEmail($user['email'], $user);
+                }
+            }
+            if (!empty($user) && $user->email && !$this->sendEmail($user['email'], $user)) {
+                $this->unassignPhoto($user);
+                DB::rollBack();
+                return 'FALSE EMAIL';
+            }
             DB::commit();
             return $user;
         } catch (\Exception $ex) {
+            $this->unassignPhoto($user);
             DB::rollBack();
             Log::info($ex->getLine());
             Log::info($ex->getMessage());
@@ -48,10 +82,16 @@ class UserService
         }
     }
 
+
     public function updateUser(array $data, $id)
     {
         try {
             $user = $this->userRepository->update($id, $data);
+            if (isset($data['role'])) {
+                $user->roles()->detach();
+                $role = $this->roleRepository->findById($data['role']);
+                $user->assignRole($role);
+            }
             DB::commit();
             return $user;
         } catch (\Exception $ex) {
@@ -65,7 +105,10 @@ class UserService
     public function deleteUser($id)
     {
         try {
-            $this->userRepository->delete($id);
+            $currentUser = User::find($id);
+            if ($this->fileService->deleteFile(cleanStorageUrl($currentUser->photo, '/storage_user/'), 'disk_user')) {
+                $this->userRepository->delete($id);
+            }
             return true;
         } catch (\Exception $ex) {
             Log::info($ex->getLine());
@@ -83,5 +126,85 @@ class UserService
             Log::info($ex->getMessage());
             throw $ex;
         }
+    }
+
+    private function assignedPhoto(&$user, $photos)
+    {
+        if (isset($photos[0])) {
+            $user->photo = $this->fileService->saveFile($photos[0], 'photo', 'disk_user');
+        }
+    }
+
+    private function unassignPhoto($user)
+    {
+        $this->fileService->deleteFile(cleanStorageUrl($user->photo, '/storage_user/'), 'disk_user');
+    }
+
+    private function sendEmail($to, $details)
+    {
+        try {
+            Mail::to($to)->send(new UserMail($details));
+            return true;
+        } catch (\Exception $ex) {
+            Log::debug($ex->getMessage());
+            return false;
+        }
+    }
+
+    public function addPhotoUser($data)
+    {
+        $flagColumn = null;
+        $user = $this->userRepository->findById($data['user_id']);
+        $columns = $this->listPhotos;
+        foreach ($columns as $column) {
+            if (empty($user->{$column})) {
+                $user->{$column} = $this->fileService->saveFile($data['photo'], 'photo', 'disk_user');
+                $flagColumn = $user->{$column};
+                $user->save();
+                break;
+            }
+        }
+        return $flagColumn;
+    }
+
+    public function deletePhotoUser($data)
+    {
+        try {
+            $matchingColumn = $this->findPhotoColumn($data['user_id'], $data['photo']);
+            if ($matchingColumn) {
+                if ($this->fileService->deleteFile(cleanStorageUrl($matchingColumn[1][$matchingColumn[0]], '/storage_user/'), 'disk_user')) {
+                    User::where('id', $data['user_id'])->update([
+                        $matchingColumn[0] => null
+                    ]);
+                }
+            }
+            return true;
+        } catch (\Exception $ex) {
+            Log::info($ex->getLine());
+            Log::info($ex->getMessage());
+            throw $ex;
+        }
+    }
+
+    private function findPhotoColumn($id, $photo)
+    {
+        $filename = basename($photo);
+        $user = User::where('id', $id)
+            ->where(function ($query) use ($filename) {
+                $query->where('photo', 'LIKE', "%$filename%");
+            })
+            ->first();
+
+        if (!$user) {
+            return null;
+        }
+
+        $columns = $this->listPhotos;
+        foreach ($columns as $column) {
+            if ($user->$column == $photo) {
+                return [$column, $user->toArray()];
+            }
+        }
+        return null;
     }
 }
